chore: separate data copy script
