+++ apps/backend/package.json
@@ -10,7 +10,7 @@
         "lint": "eslint --fix src",
         "generate": "drizzle-kit generate",
         "migrate": "tsx scripts/migrate.ts",
-        "migrate-and-copy-data": "COPY_TO_POSTGRES=true tsx scripts/migrate.ts"
+        "copy-to-pg": "tsx scripts/copy-to-pg.ts"
     },
     "dependencies": {
         "@packages/antalmanac-types": "workspace:**",
